backend/pyproject.toml
[project]
name = "slm-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.133.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.7.3",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
